[Bronze III] Title: 저작권, Time: 220 ms, Memory: 18012 KB -BaekjoonHub

+++ 백준/Bronze/2914. 저작권/저작권.java
@@ -4,7 +4,7 @@ import java.util.*;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        double a = sc.nextInt();
+        double a = sc.nextDouble();
         double l = sc.nextDouble();
         int melody = 0;
